+++ src/api/auth/auth.post.api.ts
@@ -11,6 +11,7 @@ import {
   IPhoneAuth
 } from '../types/auth';
 
+
 /* 회원가입 */
 
 export const signup = async ({
